requirements.txt
langchain==0.3.26
langchain-openai==0.3.27
openai==1.93.0
langchain-google-genai==2.1.6
faiss-cpu==1.11.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
numpy==2.1.0
tiktoken==0.9.0